infra/cloud_function/requirements.txt
google-cloud-bigquery>=2.0.0,<4.0.0
